pi-setup/requirements.txt
awscrt>=0.16.0
awsiotsdk>=1.11.0
PyGObject>=3.42.0
requests>=2.28.0
dbus-python>=1.2.18
